XRayAPI/requirements.txt
matplotlib==3.8.2
numpy==1.26.2
Pillow==10.1.0
requests==2.31.0
scipy==1.11.4
torch==2.1.1
torchvision==0.16.1
Werkzeug==3.0.1
opencv-python==4.9.0.80